:wrench: decoupled imageloader from ArtFromUrl.kt and refactors
* This will help in previews
