+++ app/src/main/java/com/shub39/rush/app/RushApp.kt
@@ -8,6 +8,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.widthIn
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
@@ -17,6 +18,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import coil3.ImageLoader
 import com.shub39.rush.billing.PaywallPage
 import com.shub39.rush.core.data.listener.NotificationListener
 import com.shub39.rush.core.presentation.RushTheme
@@ -32,31 +34,35 @@ import com.shub39.rush.viewmodels.SavedVM
 import com.shub39.rush.viewmodels.SearchSheetVM
 import com.shub39.rush.viewmodels.SettingsVM
 import com.shub39.rush.viewmodels.ShareVM
+import com.skydoves.landscapist.coil3.LocalCoilImageLoader
 import kotlinx.serialization.Serializable
 import org.koin.androidx.compose.koinViewModel
+import org.koin.compose.koinInject
 
 @Serializable
 sealed interface Route {
     @Serializable
-    data object SavedPage: Route
+    data object SavedPage : Route
 
     @Serializable
-    data object LyricsGraph: Route
+    data object LyricsGraph : Route
 
     @Serializable
-    data object SettingsGraph: Route
+    data object SettingsGraph : Route
 
     @Serializable
-    data object Onboarding: Route
+    data object Onboarding : Route
 
     @Serializable
-    data object SharePage: Route
+    data object SharePage : Route
 }
 
 @Composable
 fun RushApp() {
     val settingsVM: SettingsVM = koinViewModel()
     val searchSheetVM: SearchSheetVM = koinViewModel()
+
+    val imageLoader: ImageLoader = koinInject()
 
     val settingsState by settingsVM.state.collectAsStateWithLifecycle()
     val searchState by searchSheetVM.state.collectAsStateWithLifecycle()
@@ -70,104 +76,118 @@ fun RushApp() {
         }
     }
 
-    RushTheme(
-        state = settingsState.theme
+    CompositionLocalProvider(
+        LocalCoilImageLoader provides imageLoader
     ) {
-        AnimatedContent(
-            targetState = settingsState.showPaywall,
-            modifier = Modifier
-                .fillMaxSize()
-                .background(MaterialTheme.colorScheme.background)
+        RushTheme(
+            state = settingsState.theme
         ) {
-            if (!it) {
-                NavHost(
-                    enterTransition = { fadeIn() },
-                    exitTransition = { fadeOut() },
-                    popEnterTransition = { fadeIn() },
-                    popExitTransition = { fadeOut() },
-                    navController = navController,
-                    startDestination = Route.SavedPage,
-                    modifier = Modifier
-                        .background(MaterialTheme.colorScheme.background)
-                        .fillMaxSize()
-                ) {
-                    composable<Route.SavedPage> {
-                        val savedVM: SavedVM = koinViewModel()
-                        val savedState by savedVM.state.collectAsStateWithLifecycle()
+            AnimatedContent(
+                targetState = settingsState.showPaywall,
+                modifier = Modifier
+                    .fillMaxSize()
+                    .background(MaterialTheme.colorScheme.background)
+            ) {
+                if (!it) {
+                    NavHost(
+                        enterTransition = { fadeIn() },
+                        exitTransition = { fadeOut() },
+                        popEnterTransition = { fadeIn() },
+                        popExitTransition = { fadeOut() },
+                        navController = navController,
+                        startDestination = Route.SavedPage,
+                        modifier = Modifier
+                            .background(MaterialTheme.colorScheme.background)
+                            .fillMaxSize()
+                    ) {
+                        composable<Route.SavedPage> {
+                            val savedVM: SavedVM = koinViewModel()
+                            val savedState by savedVM.state.collectAsStateWithLifecycle()
 
-                        SavedPage(
-                            state = savedState,
-                            notificationAccess = NotificationListener.canAccessNotifications(context),
-                            action = savedVM::onAction,
-                            onNavigateToLyrics = { navController.navigate(Route.LyricsGraph) },
-                            onNavigateToSettings = { navController.navigate(Route.SettingsGraph) },
-                            modifier = Modifier.widthIn(max = 700.dp)
-                        )
+                            SavedPage(
+                                state = savedState,
+                                notificationAccess = NotificationListener.canAccessNotifications(
+                                    context
+                                ),
+                                action = savedVM::onAction,
+                                onNavigateToLyrics = { navController.navigate(Route.LyricsGraph) },
+                                onNavigateToSettings = { navController.navigate(Route.SettingsGraph) },
+                                modifier = Modifier.widthIn(max = 700.dp)
+                            )
+                        }
+
+                        composable<Route.LyricsGraph> {
+                            val lyricsVM: LyricsVM = koinViewModel()
+                            val lyricsState by lyricsVM.state.collectAsStateWithLifecycle()
+
+                            LyricsGraph(
+                                notificationAccess = NotificationListener.canAccessNotifications(
+                                    context
+                                ),
+                                lyricsState = lyricsState,
+                                lyricsAction = lyricsVM::onAction,
+                                onDismiss = { navController.navigateUp() },
+                                onShare = {
+                                    navController.navigate(Route.SharePage) {
+                                        launchSingleTop = true
+                                    }
+                                }
+                            )
+                        }
+
+                        composable<Route.SharePage> {
+                            val shareVM: ShareVM = koinViewModel()
+                            val shareState by shareVM.state.collectAsStateWithLifecycle()
+
+                            SharePage(
+                                onDismiss = { navController.navigateUp() },
+                                state = shareState,
+                                onAction = shareVM::onAction
+                            )
+                        }
+
+                        composable<Route.SettingsGraph> {
+                            SettingsGraph(
+                                notificationAccess = NotificationListener.canAccessNotifications(
+                                    context
+                                ),
+                                state = settingsState,
+                                action = settingsVM::onAction,
+                                onNavigateBack = { navController.navigateUp() }
+                            )
+                        }
+
+                        composable<Route.Onboarding> {
+                            Onboarding(
+                                onDone = {
+                                    settingsVM.onAction(
+                                        SettingsPageAction.OnUpdateOnBoardingDone(
+                                            true
+                                        )
+                                    )
+                                    navController.navigateUp()
+                                }
+                            )
+                        }
                     }
 
-                    composable<Route.LyricsGraph> {
-                        val lyricsVM: LyricsVM = koinViewModel()
-                        val lyricsState by lyricsVM.state.collectAsStateWithLifecycle()
-
-                        LyricsGraph(
-                            notificationAccess = NotificationListener.canAccessNotifications(context),
-                            lyricsState = lyricsState,
-                            lyricsAction = lyricsVM::onAction,
-                            onDismiss = { navController.navigateUp() },
-                            onShare = {
-                                navController.navigate(Route.SharePage) {
+                    if (searchState.visible) {
+                        SearchSheet(
+                            state = searchState,
+                            action = searchSheetVM::onAction,
+                            onClick = {
+                                navController.navigate(Route.LyricsGraph) {
                                     launchSingleTop = true
                                 }
                             }
                         )
                     }
-
-                    composable<Route.SharePage> {
-                        val shareVM: ShareVM = koinViewModel()
-                        val shareState by shareVM.state.collectAsStateWithLifecycle()
-
-                        SharePage(
-                            onDismiss = { navController.navigateUp() },
-                            state = shareState,
-                            onAction = shareVM::onAction
-                        )
-                    }
-
-                    composable<Route.SettingsGraph> {
-                        SettingsGraph(
-                            notificationAccess = NotificationListener.canAccessNotifications(context),
-                            state = settingsState,
-                            action = settingsVM::onAction,
-                            onNavigateBack = { navController.navigateUp() }
-                        )
-                    }
-
-                    composable<Route.Onboarding> {
-                        Onboarding(
-                            onDone = {
-                                settingsVM.onAction(SettingsPageAction.OnUpdateOnBoardingDone(true))
-                                navController.navigateUp()
-                            }
-                        )
-                    }
-                }
-
-                if (searchState.visible) {
-                    SearchSheet(
-                        state = searchState,
-                        action = searchSheetVM::onAction,
-                        onClick = {
-                            navController.navigate(Route.LyricsGraph) {
-                                launchSingleTop = true
-                            }
-                        }
+                } else {
+                    PaywallPage(
+                        isProUser = settingsState.isProUser,
+                        onDismissRequest = { settingsVM.onAction(SettingsPageAction.OnDismissPaywall) }
                     )
                 }
-            } else {
-                PaywallPage(
-                    isProUser = settingsState.isProUser,
-                    onDismissRequest = { settingsVM.onAction(SettingsPageAction.OnDismissPaywall) }
-                )
             }
         }
     }
