+++ app/src/main/java/com/shub39/rush/core/presentation/ArtFromUrl.kt
@@ -8,7 +8,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
-import coil3.ImageLoader
 import com.shub39.rush.R
 import com.skydoves.landscapist.ImageOptions
 import com.skydoves.landscapist.coil3.CoilImage
@@ -18,7 +17,6 @@ import com.skydoves.landscapist.placeholder.shimmer.ShimmerPlugin
 import compose.icons.FontAwesomeIcons
 import compose.icons.fontawesomeicons.Solid
 import compose.icons.fontawesomeicons.solid.Music
-import org.koin.compose.koinInject
 
 // General Image Composable
 @Composable
@@ -28,12 +26,10 @@ fun ArtFromUrl(
     contentScale: ContentScale = ContentScale.Crop,
     baseColor: Color = MaterialTheme.colorScheme.surface,
     highlightColor: Color = MaterialTheme.colorScheme.primary,
-    imageLoader: ImageLoader = koinInject()
 ) {
     CoilImage(
         imageModel = { imageUrl },
         modifier = modifier,
-        imageLoader = { imageLoader },
         component = rememberImageComponent {
             +ShimmerPlugin(
                 Shimmer.Resonate(
